декомпозировала код
