+++ src/components/helpers.tsx
@@ -1,4 +1,5 @@
 import { Card } from 'react-bootstrap';
+import { Dispatch, SetStateAction } from 'react';
 
 export function renderCardText(title: string, description: string) {
   return (
@@ -9,4 +10,61 @@ export function renderCardText(title: string, description: string) {
       </span>
     </Card.Text>
   );
+}
+
+//Показать еще
+export function showMoreRender(
+  description: string,
+  length: number,
+  isDetailsOpen: boolean,
+  setIsDetailsOpen: Dispatch<SetStateAction<boolean>>
+): JSX.Element {
+  const new_description = description.slice(0, length) + '...';
+  return (
+    <>
+      {description.length < length ? (
+        <p>{description}</p>
+      ) : (
+        <>
+          <p style={isDetailsOpen ? { display: 'none' } : {}}>
+            {new_description}
+          </p>
+          <p style={!isDetailsOpen ? { display: 'none' } : {}}>{description}</p>
+          <p
+            onClick={() => setIsDetailsOpen((prev) => !prev)}
+            className="mb-0 text-primary cursor-pointer fs-base"
+            style={{ fontWeight: '500' }}
+          >
+            {isDetailsOpen ? 'Свернуть' : 'Показать еще'}
+          </p>
+        </>
+      )}
+    </>
+  );
+}
+
+type SlidesCountProps = {
+  currentSlide: number;
+  totalSlides: number;
+};
+
+export const SlidesCount = ({
+  currentSlide,
+  totalSlides,
+}: SlidesCountProps) => (
+  <div className="swiper-slides-count text-dark bg-light rounded-2 p-1">
+    <i className="fi-image fs-lg me-2"></i>
+    <div className="fs-6 fw-bold ps-1 me-2">
+      <span>{currentSlide}</span>
+      <span>/</span>
+      <span>{totalSlides}</span>
+    </div>
+  </div>
+);
+
+// Единый Вид Названия
+export function capitalize(str: string) {
+  return str.replace(/(^|\s)\S/g, function (a) {
+    return a.toUpperCase();
+  });
 }
